feature: Add app discovery support for Linux
* Add new linux.rs, with discover_all() and from_app_directory() methods with implementations
  specific to Linux.
* Fill in OperatingSystem::Linux cases in file lookup methods on WolframApp
* Add conversion of 'aarch64-unknown-linux-gnu' target triple to 'Linux-ARM64' System ID

App discovery on Linux is more difficult and finicky than on macOS and Windows because
there is no standard file/format for application metadata (.plist on macOS, the registry
on Windows), so any information about an app has to be parsed out of whatever files are
in the app directory.

Undoubtedly, this initial implementation of Linux app discovery can be improved. However,
it suffices to discover and run all of the various file lookup methods for three
app/versions I tested it against:

  - Mathematica 13.1
  - Wolfram Engine v13.0
  - Wolfram Engine v13.3 (internal build)