polish: Use `log` crate to add logging output to success and error paths
This makes wolfram-app-discovery easier to debug, particular in CI/CD
environments where it can otherwise we somewhat tedious to get
useful debugging output.

There is a loose naming convention for the added logging statements:

  1. "discovery error: ..." -- emitted by every Error::*() constructor function
  2. "discovery failed: ..." -- emitted on error return paths from top-level discovery functions
  3. "discovered in (app|env): ..." -- emitted on success return paths from top-level discovery functions

(1) is commonly but not always a fatal error. For example, when using discover_all(),
    an error discovering one app does not cause the overall discovery of installed apps
    to fail.

(2) is typically a fatal error -- a top-level discovery function was forced to
    return an error because the highest-precedence discovery mechanism found an
    app or resource that produced an error during validation, or all discovery
    mechanisms were exhausted resulting in no discovery.

Note that logging output will NOT be generated unless a `log` crate logger
has been initialized by other code in the binary wolfram-app-discovery is
compiled into.

E.g. for build scripts using wolfram-app-discovery, a prerequisite for logging
output would be a call to e.g. `env_logger::init()` at the beginning of the
build script.